+++ src/com/amplitude/api/AmplitudeClient.java
@@ -852,10 +852,8 @@ public class AmplitudeClient {
             apiProperties.put("gps_enabled", deviceInfo.isGooglePlayServicesEnabled());
 
             event.put("api_properties", apiProperties);
-            event.put("event_properties", (eventProperties == null) ? new JSONObject()
-                    : truncate(eventProperties));
-            event.put("user_properties", (userProperties == null) ? new JSONObject()
-                    : truncate(userProperties));
+            event.put("event_properties", replaceWithJSONNull(truncate(eventProperties)));
+            event.put("user_properties", replaceWithJSONNull(truncate(userProperties)));
             event.put("groups", (groups == null) ? new JSONObject() : truncate(groups));
         } catch (JSONException e) {
             logger.e(TAG, e.toString());
@@ -1257,12 +1255,18 @@ public class AmplitudeClient {
                     return; // could not create copy
                 }
 
+                // sanitize and truncate properties before trying to convert to identify
+                JSONObject sanitized = truncate(copy);
+                if (sanitized == null) {
+                    return;
+                }
+
                 Identify identify = new Identify();
-                Iterator<?> keys = copy.keys();
+                Iterator<?> keys = sanitized.keys();
                 while (keys.hasNext()) {
                     String key = (String) keys.next();
                     try {
-                        identify.setUserProperty(key, copy.get(key));
+                        identify.setUserProperty(key, sanitized.get(key));
                     } catch (JSONException e) {
                         logger.e(TAG, e.toString());
                     }
@@ -1328,12 +1332,18 @@ public class AmplitudeClient {
     /**
      * Truncate values in a JSON object. Any string values longer than 1024 characters will be
      * truncated to 1024 characters.
+     * Any dictionary with more than 1000 items will be ignored.
      *
      * @param object the object
      * @return the truncated JSON object
      */
     public JSONObject truncate(JSONObject object) {
         if (object == null) {
+            return null;
+        }
+
+        if (object.length() > Constants.MAX_PROPERTY_KEYS) {
+            logger.w(TAG, "Warning: too many properties (more than 1000), ignoring");
             return null;
         }